Deprecated DataSourceClasspathDecorator
Some DataSources need to have loaded in the classpath the JDBC driver class used to connect to the database at the time of initialization. For this reason, it is not possible to delay the class loading and do it later using a decorator.

A utility class will allow applications to load classes at the appropriate time.